requirements.txt
fastapi
uvicorn
sqlalchemy[asyncio]
asyncpg
alembic
pydantic[email]
pydantic-settings
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]
